Improved cross-language homogeneity (#34)
Use an explicit "handle" method on Function instances, as this is the
most similar method signature approach between different language; a
primary goal of the project.  This is in contrast to a more strict
adherence to the ASGI spec which would see the Function object itself be
callable.  See PR for further exploration of this choice.